requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
outcome==1.1.0
psutil==5.9.1
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.1
selenium==4.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
whitenoise==6.1.0
wsproto==1.1.0
